fixed uart baudrate issue

+++ src/uart.c
@@ -8,7 +8,7 @@ uint8_t UART_BUFFER[UART_BUFFER_LENGTH];
 uint8_t uartLastRead = 0;
 uint8_t uartLastRecevied = 0;
 
-void uart_init(uint8_t baudrate) {
+void uart_init(uint16_t baudrate) {
     // Set baud rate
     UBRR0H = (uint8_t)(baudrate >> 8);
     UBRR0L = (uint8_t)baudrate;
