+++ src/uart.h
@@ -7,7 +7,7 @@
 
 #include <stdint.h>
 
-void uart_init(uint8_t baudrate);
+void uart_init(uint16_t baudrate);
 void uart_transmit(uint8_t data);
 uint8_t uart_receive(void);
